+++ src/cv2ext/datasets/__init__.py
@@ -1,0 +1,18 @@
+# Copyright (c) 2024 Justin Davis (davisjustin302@gmail.com)
+#
+# MIT License
+"""
+Submodule containing tools for reading and interacting with datasets.
+
+Submodules
+----------
+:mod:`mot`
+    Tools for interacting with MOT17(det)/MOT20(det) datasets.
+
+"""
+
+from __future__ import annotations
+
+from . import mot
+
+__all__ = ["mot"]